Woop. Passwords and everything

+++ deployer.sh
@@ -1,12 +1,15 @@
 #!/bin/bash -
 
-cd $1
+DIR=$1
+PASSWORD=$2
+NAME=meteor-book
 
 FIXTURES=server/fixtures.js
 RESET=server/reset.js
 
 function prepare_reset {
-  rm $RESET
+  [ -e $RESET ] && rm $RESET
+  
   if [ -e $FIXTURES ]
   then
     cat > $RESET <<JAVASCRIPT
@@ -23,17 +26,18 @@ JAVASCRIPT
     
     cat >> $RESET <<JAVASCRIPT
       
-      }, 24 * 3600 * 1000); // every day
+      }, 1000); // every day
 JAVASCRIPT
     
     cat $RESET
   fi
 }
 
-
+cd $DIR
 for tag in `git tag`
 do
   git reset --hard $tag
   prepare_reset
-  mrt deploy meteor-book-$tag.meteor.com
+  echo "$PASSWORD
+$PASSWORD" | mrt deploy $NAME-$tag -P
 done
